hotfix

+++ db/db.sql
@@ -1382,7 +1382,7 @@ ALTER TABLE promocodes
 ADD UNIQUE INDEX code (code)   ; 
 
 CREATE TABLE custacc (
-  ca_id int(11) NOT NULL AUTO_INCREMENT,
+  ca_id bigint(20) NOT NULL AUTO_INCREMENT,
   customer_id int(11) NOT NULL,
   document_id int(11) DEFAULT NULL,
   optype int(11) DEFAULT NULL,
